[➕] Add script bash on the root to launch and canceled all

+++ script.sh
@@ -1,16 +1,41 @@
 #!/bin/bash
 
-cd Backend 
-bash script.sh full
-echo "[Server BACKEND] is launch"
 
-cd ../Frontend
-docker-compose up --build
-echo "[Server FRONTEND] is launch"
+cleanup() {
 
+    echo -n "Do you want to shut down all servers? (Y/n): "
+    read -r response
+
+    response=$(echo "$response")
+
+    if [[ -z "$response" || "$response" == "y" ]]; then
+        echo "[WAIT -------------------------- SHUTTING DOWN]"
+
+        echo "[SHUTTING DOWN BACKEND]"
+        cd Backend
+        docker-compose down
+        cd ..
+
+        echo "[SHUTTING DOWN FRONTEND]"
+        cd Frontend
+        docker-compose down
+        cd ..
+
+        echo "[All SERVERS ARE SHUTTING DOWN]"
+        exit 0
+    else
+        echo "Shutdown canceled. Servers are still running."
+    fi
+}
 
 trap cleanup SIGINT
-echo "[SHUT DOWN BACKEND DOCKER]"
 
-cd Backend
-docker-compose down
+cd Backend 
+bash script.sh -full &
+echo "[Server BACKEND] is running"
+
+cd ../Frontend
+docker-compose up --build &
+echo "[Server FRONTEND] is running"
+
+wait
